+++ packages/components/Callout/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/callout
 
-This log was last generated on Wed, 04 May 2022 21:06:04 GMT and should not be manually modified.
+This log was last generated on Tue, 10 May 2022 20:49:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.20.0
+
+Tue, 10 May 2022 20:49:25 GMT
+
+### Minor changes
+
+- Add doNotTakePointerCapture prop to Callout interface (patboyd@microsoft.com)
 
 ## 0.19.46