+++ packages/experimental/Menu/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/menu
 
-This log was last generated on Wed, 04 May 2022 21:06:04 GMT and should not be manually modified.
+This log was last generated on Wed, 11 May 2022 22:26:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.0
+
+Wed, 11 May 2022 22:26:25 GMT
+
+### Minor changes
+
+- Dismiss and openOnHover (ruaraki@microsoft.com)
+
+## 0.5.2
+
+Wed, 11 May 2022 19:33:59 GMT
+
+### Patches
+
+- Add MenuItemCheckbox (ruaraki@microsoft.com)
+
+## 0.5.1
+
+Tue, 10 May 2022 20:49:25 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/callout to v0.20.0
 
 ## 0.5.0